refactor(api): split hierarchy-service-factory by responsibility
Bean: api-u109
