+++ apps/api/src/services/hierarchy-service-factory.ts
@@ -1,501 +1,7 @@
-import { systems } from "@pluralscape/db/pg";
-import { createId, now } from "@pluralscape/types";
-import { and, count, eq, gt, sql } from "drizzle-orm";
-
-import { HTTP_CONFLICT, HTTP_NOT_FOUND, HTTP_TOO_MANY_REQUESTS } from "../http.constants.js";
-import { ApiHttpError } from "../lib/api-error.js";
-import { checkDependents } from "../lib/check-dependents.js";
-import { validateEncryptedBlob } from "../lib/encrypted-blob.js";
-import { archiveEntity } from "../lib/entity-lifecycle.js";
-import { assertOccUpdated } from "../lib/occ-update.js";
-import { buildPaginatedResult } from "../lib/pagination.js";
-import { withTenantRead, withTenantTransaction } from "../lib/rls-context.js";
-import { assertSystemOwnership } from "../lib/system-ownership.js";
-import { tenantCtx } from "../lib/tenant-context.js";
-import {
-  DEFAULT_PAGE_LIMIT,
-  MAX_ENCRYPTED_DATA_BYTES,
-  MAX_PAGE_LIMIT,
-} from "../service.constants.js";
-
-import { dispatchWebhookEvent } from "./webhook-dispatcher.js";
-
-import type {
-  HierarchyCreateBody,
-  HierarchyService,
-  HierarchyServiceConfig,
-  HierarchyUpdateBody,
-} from "./hierarchy-service-types.js";
-import type { AuditWriter } from "../lib/audit-writer.js";
-import type { AuthContext } from "../lib/auth-context.js";
-import type { ArchivableEntityConfig } from "../lib/entity-lifecycle.js";
-import type { PaginatedResult, SystemId } from "@pluralscape/types";
-import type { PostgresJsDatabase } from "drizzle-orm/postgres-js";
-
-// ── Factory ────────────────────────────────────────────────────────
-
-/** Create a hierarchy service with standard CRUD, archive, and restore operations. */
-export function createHierarchyService<
-  TRow extends Record<string, unknown>,
-  TId extends string,
-  TResult extends { readonly id: string },
-  TCreateBody extends HierarchyCreateBody,
-  TUpdateBody extends HierarchyUpdateBody,
->(
-  cfg: HierarchyServiceConfig<TRow, TResult, TCreateBody, TUpdateBody>,
-): HierarchyService<TId, TResult, TCreateBody, TUpdateBody> {
-  const {
-    table,
-    columns,
-    idPrefix,
-    entityName,
-    parentFieldName,
-    toResult,
-    events,
-    dependentChecks,
-  } = cfg;
-
-  const lifecycleCfg: ArchivableEntityConfig<TId> = {
-    table,
-    columns: {
-      id: columns.id,
-      systemId: columns.systemId,
-      archived: columns.archived,
-      archivedAt: columns.archivedAt,
-      updatedAt: columns.updatedAt,
-      version: columns.version,
-    },
-    entityName,
-    archiveEvent: events.archived,
-    restoreEvent: events.restored,
-  };
-
-  // ── CREATE ────────────────────────────────────────────────────────
-
-  async function create(
-    db: PostgresJsDatabase,
-    systemId: SystemId,
-    body: TCreateBody,
-    auth: AuthContext,
-    audit: AuditWriter,
-  ): Promise<TResult> {
-    assertSystemOwnership(systemId, auth);
-
-    const blob = validateEncryptedBlob(body.encryptedData, MAX_ENCRYPTED_DATA_BYTES);
-
-    const entityId = createId(idPrefix);
-    const timestamp = now();
-
-    return withTenantTransaction(db, tenantCtx(systemId, auth), async (tx) => {
-      // Enforce per-system quota if configured
-      if (cfg.maxPerSystem !== undefined) {
-        await tx
-          .select({ id: systems.id })
-          .from(systems)
-          .where(eq(systems.id, systemId))
-          .for("update");
-
-        const [existing] = await tx
-          .select({ count: count() })
-          .from(table)
-          .where(and(eq(columns.systemId, systemId), eq(columns.archived, false)));
-
-        if ((existing?.count ?? 0) >= cfg.maxPerSystem) {
-          throw new ApiHttpError(
-            HTTP_TOO_MANY_REQUESTS,
-            "QUOTA_EXCEEDED",
-            `Maximum of ${String(cfg.maxPerSystem)} ${entityName.toLowerCase()}s per system`,
-          );
-        }
-      }
-
-      // Validate parent exists in same system if non-null. The body is typed
-      // generically — the parent field is referenced by name from cfg.parentFieldName,
-      // so we read the value via Reflect.get which keeps strict typing without
-      // any cast.
-      const rawParentId: unknown = Reflect.get(body, parentFieldName);
-      const parentId = typeof rawParentId === "string" ? rawParentId : null;
-      if (parentId !== null) {
-        const [parent] = await tx
-          .select({ id: columns.id })
-          .from(table)
-          .where(
-            and(
-              eq(columns.id, parentId),
-              eq(columns.systemId, systemId),
-              eq(columns.archived, false),
-            ),
-          )
-          .limit(1);
-
-        if (!parent) {
-          throw new ApiHttpError(
-            HTTP_NOT_FOUND,
-            "NOT_FOUND",
-            `Parent ${entityName.toLowerCase()} not found`,
-          );
-        }
-      }
-
-      const extraValues = cfg.createInsertValues(body);
-
-      const [row] = await tx
-        .insert(table)
-        .values({
-          id: entityId,
-          systemId,
-          [parentFieldName]: parentId ?? null,
-          encryptedData: blob,
-          createdAt: timestamp,
-          updatedAt: timestamp,
-          ...extraValues,
-        } as Record<string, unknown>)
-        .returning();
-
-      if (!row) {
-        throw new Error(`Failed to create ${entityName.toLowerCase()} — INSERT returned no rows`);
-      }
-
-      await audit(tx, {
-        eventType: events.created,
-        actor: { kind: "account", id: auth.accountId },
-        detail: `${entityName} created`,
-        systemId,
-      });
-      if (cfg.webhookEvents) {
-        await dispatchWebhookEvent(
-          tx,
-          systemId,
-          cfg.webhookEvents.created,
-          cfg.webhookEvents.buildPayload(entityId),
-        );
-      }
-
-      return toResult(row as TRow);
-    });
-  }
-
-  // ── LIST ──────────────────────────────────────────────────────────
-
-  async function list(
-    db: PostgresJsDatabase,
-    systemId: SystemId,
-    auth: AuthContext,
-    cursor?: string,
-    limit = DEFAULT_PAGE_LIMIT,
-    includeArchived = false,
-  ): Promise<PaginatedResult<TResult>> {
-    assertSystemOwnership(systemId, auth);
-
-    const effectiveLimit = Math.min(limit, MAX_PAGE_LIMIT);
-
-    return withTenantRead(db, tenantCtx(systemId, auth), async (tx) => {
-      const conditions = [eq(columns.systemId, systemId)];
-
-      if (!includeArchived) {
-        conditions.push(eq(columns.archived, false));
-      }
-
-      if (cursor) {
-        conditions.push(gt(columns.id, cursor));
-      }
-
-      const rows = await tx
-        .select()
-        .from(table)
-        .where(and(...conditions))
-        .orderBy(columns.id)
-        .limit(effectiveLimit + 1);
-
-      return buildPaginatedResult(rows, effectiveLimit, (row) => toResult(row as TRow));
-    });
-  }
-
-  // ── GET ───────────────────────────────────────────────────────────
-
-  async function get(
-    db: PostgresJsDatabase,
-    systemId: SystemId,
-    entityId: TId,
-    auth: AuthContext,
-  ): Promise<TResult> {
-    assertSystemOwnership(systemId, auth);
-
-    return withTenantRead(db, tenantCtx(systemId, auth), async (tx) => {
-      const [row] = await tx
-        .select()
-        .from(table)
-        .where(
-          and(
-            eq(columns.id, entityId),
-            eq(columns.systemId, systemId),
-            eq(columns.archived, false),
-          ),
-        )
-        .limit(1);
-
-      if (!row) {
-        throw new ApiHttpError(HTTP_NOT_FOUND, "NOT_FOUND", `${entityName} not found`);
-      }
-
-      return toResult(row as TRow);
-    });
-  }
-
-  // ── UPDATE ────────────────────────────────────────────────────────
-
-  async function update(
-    db: PostgresJsDatabase,
-    systemId: SystemId,
-    entityId: TId,
-    body: TUpdateBody,
-    auth: AuthContext,
-    audit: AuditWriter,
-  ): Promise<TResult> {
-    assertSystemOwnership(systemId, auth);
-
-    const blob = validateEncryptedBlob(body.encryptedData, MAX_ENCRYPTED_DATA_BYTES);
-
-    const timestamp = now();
-
-    return withTenantTransaction(db, tenantCtx(systemId, auth), async (tx) => {
-      if (cfg.beforeUpdate) {
-        await cfg.beforeUpdate(tx, entityId, body, systemId);
-      }
-
-      const extraValues = cfg.updateSetValues(body);
-
-      const updated = await tx
-        .update(table)
-        .set({
-          encryptedData: blob,
-          updatedAt: timestamp,
-          version: sql`${columns.version} + 1`,
-          ...extraValues,
-        } as Record<string, unknown>)
-        .where(
-          and(
-            eq(columns.id, entityId),
-            eq(columns.systemId, systemId),
-            eq(columns.version, body.version),
-            eq(columns.archived, false),
-          ),
-        )
-        .returning();
-
-      const row = await assertOccUpdated(
-        updated,
-        async () => {
-          const [existing] = await tx
-            .select({ id: columns.id })
-            .from(table)
-            .where(
-              and(
-                eq(columns.id, entityId),
-                eq(columns.systemId, systemId),
-                eq(columns.archived, false),
-              ),
-            )
-            .limit(1);
-          return existing;
-        },
-        entityName,
-      );
-
-      await audit(tx, {
-        eventType: events.updated,
-        actor: { kind: "account", id: auth.accountId },
-        detail: `${entityName} updated`,
-        systemId,
-      });
-      if (cfg.webhookEvents) {
-        await dispatchWebhookEvent(
-          tx,
-          systemId,
-          cfg.webhookEvents.updated,
-          cfg.webhookEvents.buildPayload(entityId),
-        );
-      }
-
-      return toResult(row as TRow);
-    });
-  }
-
-  // ── DELETE ────────────────────────────────────────────────────────
-
-  async function remove(
-    db: PostgresJsDatabase,
-    systemId: SystemId,
-    entityId: TId,
-    auth: AuthContext,
-    audit: AuditWriter,
-  ): Promise<void> {
-    assertSystemOwnership(systemId, auth);
-
-    await withTenantTransaction(db, tenantCtx(systemId, auth), async (tx) => {
-      // Verify entity exists
-      const [existing] = await tx
-        .select({ id: columns.id })
-        .from(table)
-        .where(
-          and(
-            eq(columns.id, entityId),
-            eq(columns.systemId, systemId),
-            eq(columns.archived, false),
-          ),
-        )
-        .limit(1);
-
-      if (!existing) {
-        throw new ApiHttpError(HTTP_NOT_FOUND, "NOT_FOUND", `${entityName} not found`);
-      }
-
-      // Translate the factory's narrow shape (entityColumn + systemColumn
-      // + optional archived filter) into the shared helper's generic
-      // predicate form. Error payload matches other services: narrative
-      // message plus a structured `{ dependents }` detail for clients that
-      // want to render per-type counts.
-      const checks = dependentChecks.map((dep) => {
-        const conditions = [eq(dep.entityColumn, entityId), eq(dep.systemColumn, systemId)];
-        if (dep.filterArchived) {
-          conditions.push(eq(dep.filterArchived, false));
-        }
-        return {
-          table: dep.table,
-          predicate: and(...conditions),
-          typeName: dep.label,
-        };
-      });
-
-      const { dependents } = await checkDependents(tx, checks);
-
-      if (dependents.length > 0) {
-        const parts = dependents.map((d) => `${String(d.count)} ${d.type}`);
-        throw new ApiHttpError(
-          HTTP_CONFLICT,
-          "HAS_DEPENDENTS",
-          `${entityName} has ${parts.join(" and ")}. Remove all dependents before deleting.`,
-          { dependents },
-        );
-      }
-
-      // Audit before delete (FK satisfied since entity still exists)
-      await audit(tx, {
-        eventType: events.deleted,
-        actor: { kind: "account", id: auth.accountId },
-        detail: `${entityName} deleted`,
-        systemId,
-      });
-
-      // Hard delete
-      await tx.delete(table).where(and(eq(columns.id, entityId), eq(columns.systemId, systemId)));
-    });
-  }
-
-  // ── ARCHIVE ───────────────────────────────────────────────────────
-
-  async function archive(
-    db: PostgresJsDatabase,
-    systemId: SystemId,
-    entityId: TId,
-    auth: AuthContext,
-    audit: AuditWriter,
-  ): Promise<void> {
-    await archiveEntity(db, systemId, entityId, auth, audit, lifecycleCfg);
-  }
-
-  // ── RESTORE ───────────────────────────────────────────────────────
-
-  async function restore(
-    db: PostgresJsDatabase,
-    systemId: SystemId,
-    entityId: TId,
-    auth: AuthContext,
-    audit: AuditWriter,
-  ): Promise<TResult> {
-    assertSystemOwnership(systemId, auth);
-
-    const timestamp = now();
-
-    return withTenantTransaction(db, tenantCtx(systemId, auth), async (tx) => {
-      const [existing] = await tx
-        .select({ id: columns.id, parentId: columns.parentId })
-        .from(table)
-        .where(
-          and(eq(columns.id, entityId), eq(columns.systemId, systemId), eq(columns.archived, true)),
-        )
-        .limit(1);
-
-      if (!existing) {
-        throw new ApiHttpError(
-          HTTP_NOT_FOUND,
-          "NOT_FOUND",
-          `Archived ${entityName.toLowerCase()} not found`,
-        );
-      }
-
-      // Enforce per-system quota on restore if configured
-      if (cfg.maxPerSystem !== undefined) {
-        const [activeCount] = await tx
-          .select({ count: count() })
-          .from(table)
-          .where(and(eq(columns.systemId, systemId), eq(columns.archived, false)));
-
-        if ((activeCount?.count ?? 0) >= cfg.maxPerSystem) {
-          throw new ApiHttpError(
-            HTTP_TOO_MANY_REQUESTS,
-            "QUOTA_EXCEEDED",
-            `Maximum of ${String(cfg.maxPerSystem)} ${entityName.toLowerCase()}s per system`,
-          );
-        }
-      }
-
-      // If parent is archived, promote to root
-      let newParentId = typeof existing.parentId === "string" ? existing.parentId : null;
-      if (newParentId !== null) {
-        const [parent] = await tx
-          .select({ archived: columns.archived })
-          .from(table)
-          .where(and(eq(columns.id, newParentId), eq(columns.systemId, systemId)))
-          .limit(1);
-
-        if (!parent || (parent.archived as boolean)) {
-          newParentId = null;
-        }
-      }
-
-      const updated = await tx
-        .update(table)
-        .set({
-          archived: false,
-          archivedAt: null,
-          [parentFieldName]: newParentId,
-          updatedAt: timestamp,
-          version: sql`${columns.version} + 1`,
-        } as Record<string, unknown>)
-        .where(and(eq(columns.id, entityId), eq(columns.systemId, systemId)))
-        .returning();
-
-      if (updated.length === 0) {
-        throw new ApiHttpError(
-          HTTP_NOT_FOUND,
-          "NOT_FOUND",
-          `Archived ${entityName.toLowerCase()} not found`,
-        );
-      }
-
-      const [row] = updated as [(typeof updated)[number], ...typeof updated];
-
-      await audit(tx, {
-        eventType: events.restored,
-        actor: { kind: "account", id: auth.accountId },
-        detail: `${entityName} restored`,
-        systemId,
-      });
-
-      return toResult(row as TRow);
-    });
-  }
-
-  return { create, list, get, update, remove, archive, restore };
-}
+/**
+ * Hierarchy service factory — public barrel.
+ *
+ * Callers import `createHierarchyService` from this path. The implementation
+ * is split by verb across `hierarchy-service-factory/` sub-modules.
+ */
+export { createHierarchyService } from "./hierarchy-service-factory/factory.js";
